Refactor shared testing code from core_test
