+++ nano/core_test/distributed_work.cpp
@@ -1,6 +1,6 @@
 #include <nano/core_test/fakes/work_peer.hpp>
-#include <nano/core_test/testutil.hpp>
 #include <nano/node/testing.hpp>
+#include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
