+++ nano/core_test/gap_cache.cpp
@@ -1,5 +1,5 @@
-#include <nano/core_test/testutil.hpp>
 #include <nano/node/testing.hpp>
+#include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
