+++ nano/core_test/ipc.cpp
@@ -1,10 +1,10 @@
-#include <nano/core_test/testutil.hpp>
 #include <nano/lib/ipc_client.hpp>
 #include <nano/lib/tomlconfig.hpp>
 #include <nano/node/ipc/ipc_access_config.hpp>
 #include <nano/node/ipc/ipc_server.hpp>
 #include <nano/node/testing.hpp>
 #include <nano/rpc/rpc.hpp>
+#include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
