+++ nano/core_test/ledger.cpp
@@ -1,8 +1,8 @@
-#include <nano/core_test/testutil.hpp>
 #include <nano/lib/stats.hpp>
 #include <nano/lib/threading.hpp>
 #include <nano/node/election.hpp>
 #include <nano/node/testing.hpp>
+#include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
