+++ nano/core_test/locks.cpp
@@ -1,5 +1,5 @@
-#include <nano/core_test/testutil.hpp>
 #include <nano/lib/locks.hpp>
+#include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
