+++ nano/core_test/message_parser.cpp
@@ -1,4 +1,5 @@
 #include <nano/node/testing.hpp>
+#include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
