+++ nano/core_test/network_filter.cpp
@@ -1,8 +1,8 @@
-#include <nano/core_test/testutil.hpp>
 #include <nano/node/common.hpp>
 #include <nano/secure/buffer.hpp>
 #include <nano/secure/common.hpp>
 #include <nano/secure/network_filter.hpp>
+#include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
