+++ nano/core_test/node_telemetry.cpp
@@ -1,7 +1,6 @@
-#include <nano/core_test/common.hpp>
-#include <nano/core_test/testutil.hpp>
 #include <nano/node/telemetry.hpp>
 #include <nano/node/testing.hpp>
+#include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
