+++ nano/core_test/processor_service.cpp
@@ -1,11 +1,11 @@
 #include <nano/boost/asio/io_context.hpp>
-#include <nano/core_test/testutil.hpp>
 #include <nano/lib/alarm.hpp>
 #include <nano/lib/stats.hpp>
 #include <nano/lib/work.hpp>
 #include <nano/secure/blockstore.hpp>
 #include <nano/secure/ledger.hpp>
 #include <nano/secure/utility.hpp>
+#include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
