+++ nano/load_test/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_executable (load_test
 	entry.cpp)
 
-target_link_libraries (load_test node secure Boost::boost)
+target_link_libraries (load_test node secure test_common gtest Boost::boost)
 
 target_compile_definitions(load_test
 	PRIVATE
