+++ nano/node/testing.cpp
@@ -1,5 +1,3 @@
-#define IGNORE_GTEST_INCL
-#include <nano/core_test/testutil.hpp>
 #include <nano/crypto_lib/random_pool.hpp>
 #include <nano/node/common.hpp>
 #include <nano/node/testing.hpp>
@@ -8,6 +6,8 @@
 #include <boost/property_tree/json_parser.hpp>
 
 #include <cstdlib>
+
+using namespace std::chrono_literals;
 
 std::string nano::error_system_messages::message (int ev) const
 {
@@ -176,19 +176,20 @@ uint64_t nano::system::work_generate_limited (nano::block_hash const & root_a, u
 std::unique_ptr<nano::state_block> nano::upgrade_epoch (nano::work_pool & pool_a, nano::ledger & ledger_a, nano::epoch epoch_a)
 {
 	auto transaction (ledger_a.store.tx_begin_write ());
-	auto account = nano::test_genesis_key.pub;
+	auto test_genesis_key = nano::ledger_constants (nano::nano_networks::nano_test_network).test_genesis_key;
+	auto account = test_genesis_key.pub;
 	auto latest = ledger_a.latest (transaction, account);
 	auto balance = ledger_a.account_balance (transaction, account);
 
 	nano::state_block_builder builder;
 	std::error_code ec;
 	auto epoch = builder
-	             .account (nano::test_genesis_key.pub)
+	             .account (test_genesis_key.pub)
 	             .previous (latest)
 	             .balance (balance)
 	             .link (ledger_a.epoch_link (epoch_a))
-	             .representative (nano::test_genesis_key.pub)
-	             .sign (nano::test_genesis_key.prv, nano::test_genesis_key.pub)
+	             .representative (test_genesis_key.pub)
+	             .sign (test_genesis_key.prv, test_genesis_key.pub)
 	             .work (*pool_a.generate (latest, nano::work_threshold (nano::work_version::work_1, nano::block_details (epoch_a, false, false, true))))
 	             .build (ec);
 
@@ -327,7 +328,7 @@ void nano::system::generate_receive (nano::node & node_a)
 	}
 	if (send_block != nullptr)
 	{
-		auto receive_error (wallet (0)->receive_sync (send_block, nano::genesis_account, std::numeric_limits<nano::uint128_t>::max ()));
+		auto receive_error (wallet (0)->receive_sync (send_block, nano::ledger_constants (nano::nano_networks::nano_test_network).genesis_account, std::numeric_limits<nano::uint128_t>::max ()));
 		(void)receive_error;
 	}
 }
@@ -452,8 +453,9 @@ void nano::system::generate_send_new (nano::node & node_a, std::vector<nano::acc
 void nano::system::generate_mass_activity (uint32_t count_a, nano::node & node_a)
 {
 	std::vector<nano::account> accounts;
-	wallet (0)->insert_adhoc (nano::test_genesis_key.prv);
-	accounts.push_back (nano::test_genesis_key.pub);
+	auto test_genesis_key = nano::ledger_constants (nano::nano_networks::nano_test_network).test_genesis_key;
+	wallet (0)->insert_adhoc (test_genesis_key.prv);
+	accounts.push_back (test_genesis_key.pub);
 	auto previous (std::chrono::steady_clock::now ());
 	for (uint32_t i (0); i < count_a; ++i)
 	{
@@ -485,9 +487,31 @@ void nano::system::stop ()
 	work.stop ();
 }
 
-namespace nano
+uint16_t nano::get_available_port ()
 {
-void cleanup_test_directories_on_exit ()
+	// Maximum possible sockets which may feasibly be used in 1 test
+	constexpr auto max = 200;
+	static uint16_t current = 0;
+	// Read the TEST_BASE_PORT environment and override the default base port if it exists
+	auto base_str = std::getenv ("TEST_BASE_PORT");
+	uint16_t base_port = 24000;
+	if (base_str)
+	{
+		base_port = boost::lexical_cast<uint16_t> (base_str);
+	}
+
+	uint16_t const available_port = base_port + current;
+	++current;
+	// Reset port number once we have reached the maximum
+	if (current == max)
+	{
+		current = 0;
+	}
+
+	return available_port;
+}
+
+void nano::cleanup_test_directories_on_exit ()
 {
 	// Makes sure everything is cleaned up
 	nano::logging::release_file_sink ();
@@ -498,5 +522,4 @@ void cleanup_test_directories_on_exit ()
 	{
 		nano::remove_temporary_directories ();
 	}
-}
 }
