+++ nano/qt_test/qt.cpp
@@ -1,6 +1,6 @@
-#include <nano/core_test/testutil.hpp>
 #include <nano/node/testing.hpp>
 #include <nano/qt/qt.hpp>
+#include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
