+++ nano/rpc_test/CMakeLists.txt
@@ -2,7 +2,7 @@ add_executable (rpc_test
 	entry.cpp
 	rpc.cpp)
 
-target_link_libraries(rpc_test gtest rpc node)
+target_link_libraries(rpc_test node rpc test_common gtest)
 
 target_compile_definitions(rpc_test
 	PUBLIC
