+++ nano/slow_test/CMakeLists.txt
@@ -2,4 +2,4 @@ add_executable (slow_test
 	entry.cpp
 	node.cpp)
 
-target_link_libraries (slow_test node secure gtest libminiupnpc-static Boost::boost)
+target_link_libraries (slow_test secure node test_common gtest libminiupnpc-static Boost::boost)
