+++ nano/slow_test/node.cpp
@@ -1,10 +1,9 @@
-#include <nano/core_test/common.hpp>
-#include <nano/core_test/testutil.hpp>
 #include <nano/crypto_lib/random_pool.hpp>
 #include <nano/lib/threading.hpp>
 #include <nano/node/election.hpp>
 #include <nano/node/testing.hpp>
 #include <nano/node/transport/udp.hpp>
+#include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
