+++ nano/test_common/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library (test_common
+		testutil.hpp
+		testutil.cpp)
+
+target_link_libraries (test_common node gtest Boost::log_setup Boost::log Boost::boost)
